feat: added images in the readme file, modified getaccesstoken function as common in utils
