+++ pages/api/auth/passkey/authenticate.ts
@@ -1,4 +1,5 @@
 import type { NextApiRequest, NextApiResponse } from "next";
+import { getAccessToken } from "../../../../utils/utils";
 
 const {
   AUTHACTION_TENANT_DOMAIN: TENANT_DOMAIN = "",
@@ -7,36 +8,7 @@ const {
   AUTHACTION_APP_ID: APP_ID = "",
 } = process.env;
 
-let cached: { token: string; expiresAt: number } | null = null;
 
-export async function getAccessToken(): Promise<string> {
-  const now = Date.now();
-  if (cached && cached.expiresAt > now) {
-    return cached.token;
-  }
-
-  const res = await fetch(`https://${TENANT_DOMAIN}/oauth2/m2m/token`, {
-    method: "POST",
-    headers: { "Content-Type": "application/json" },
-    body: JSON.stringify({
-      client_id: CLIENT_ID,
-      client_secret: CLIENT_SECRET,
-      audience: `https://${TENANT_DOMAIN}`,
-      grant_type: "client_credentials",
-    }),
-  });
-
-  if (!res.ok) {
-    throw new Error(`M2M token fetch failed: ${res.status}`);
-  }
-
-  const { access_token, expires_in } = await res.json();
-  cached = {
-    token: access_token,
-    expiresAt: now + expires_in * 1000 - 5000,
-  };
-  return access_token;
-}
 
 export default async function handler(req: NextApiRequest, res: NextApiResponse) {
   if (req.method !== "POST") {
